TenancyClassification -> TenancyClassificationV1
Intention here is to run the old rule engine in parallel with the new
one.

This will allow us to run cases through both engines during the sync
job, and evaluate discrepancies.
