+++ spec/lib/hackney/income/tenancy_classification_v1_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+describe Hackney::Income::TenancyClassificationV1 do
+  it_behaves_like 'TenancyClassification Contract'
+end
